+++ Rio/Rio.Web/Modules/Workspace/ActivationLog/ActivationLogForm.cs
@@ -14,7 +14,7 @@ namespace Rio.Workspace.Forms
         [HalfWidth]
         public string SerialKey { get; set; }
         [HalfWidth]
-        public int TeacherId { get; set; }
+        public long TeacherId { get; set; }
         [HalfWidth]
         public int ExamListId { get; set; }
         [HalfWidth]
